+++ src/microsoft/opentelemetry/a365/core/exporters/enriched_span.py
@@ -18,6 +18,7 @@ from opentelemetry.util import types
 
 # mypy: disable-error-code="no-untyped-def"
 
+
 # pylint:disable=super-init-not-called
 class EnrichedReadableSpan(ReadableSpan):
     """Wrapper to add attributes to an immutable ReadableSpan.