+++ src/microsoft/opentelemetry/a365/core/message_utils.py
@@ -29,6 +29,7 @@ from microsoft.opentelemetry.a365.core.models.messages import (
 
 logger = logging.getLogger(__name__)
 
+
 # pylint: disable=broad-exception-caught
 def is_string_list(
     param: Union[InputMessagesParam, OutputMessagesParam],